refactor(perps): move parseVolume to component utils and align explore limit with mobile
Move parseVolume from page-level sortMarkets.ts to component-level
utils.ts to fix a layering inversion where hooks and components
imported from a page module. Also extract the explore markets limit
into PERPS_CONSTANTS.EXPLORE_MARKETS_LIMIT (8) to match mobile.
